Fix DBC DATASUS reading on some cases
The size of output buffer for uncompressed data was badly guessed as
4*size of compressed data.

Now the size is calculated from header data (number of records * size of
each record). Now it fits all cases and is not wasting space.

Note that it's not optimal and there's room for improvement. All
operations are performed in memory and final result (all unmpressed
data) is stored in a Queue<Byte>